upload/requirements.txt
bcrypt==4.0.1
Flask==3.0.0
Flask-Cors==4.0.0
openai==1.3.8
PyJWT==2.8.0
requests==2.31.0